Extract middleware into separate modules
