+++ src/lib.rs
@@ -14,8 +14,8 @@ extern crate url;
 extern crate urlencoded;
 
 use emailaddress::EmailAddress;
-use iron::headers::{ContentType, Location, StrictTransportSecurity};
-use iron::middleware::{Handler, BeforeMiddleware, AfterMiddleware};
+use iron::headers::{ContentType, Location};
+use iron::middleware::Handler;
 use iron::modifiers;
 use iron::method::Method;
 use iron::prelude::*;
@@ -34,6 +34,7 @@ pub mod config;
 pub use config::{Config, ConfigBuilder};
 pub mod crypto;
 pub mod email;
+pub mod middleware;
 pub mod oidc;
 pub mod store;
 pub mod store_cache;
@@ -42,11 +43,6 @@ use validation::{valid_uri, only_origin, same_origin};
 
 use error::{BrokerResult, BrokerError};
 
-header! { (ContentSecurityPolicy, "Content-Security-Policy") => [String] }
-header! { (XContentSecurityPolicy, "X-Content-Security-Policy") => [String] }
-header! { (XContentTypeOptions, "X-Content-Type-Options") => [String] }
-header! { (XXSSProtection, "X-XSS-Protection") => [String] }
-header! { (XFrameOptions, "X-Frame-Options") => [String] }
 
 
 /// Iron extension key we use to store the `redirect_uri`.
@@ -193,49 +189,6 @@ fn handle_error(app: &Config, req: &mut Request, err: BrokerError) -> IronResult
                                          ("error", "internal server error"),
                                      ]))))
         },
-    }
-}
-
-
-/// Middleware that logs each request.
-pub struct LogMiddleware;
-impl BeforeMiddleware for LogMiddleware {
-    fn before(&self, req: &mut Request) -> IronResult<()> {
-        info!("{} {}", req.method, req.url);
-        Ok(())
-    }
-}
-
-
-/// Middleware that sets common headers.
-pub struct DefaultHeadersMiddleware;
-impl DefaultHeadersMiddleware {
-    fn set_headers(&self, res: &mut Response) {
-        // Specify a tight content security policy. We need to be able to POST
-        // redirect anywhere, and run our own inline scripts.
-        let csp = vec![
-            "sandbox allow-scripts allow-forms",
-            "default-src 'none'",
-            "script-src 'self'",
-            "style-src 'self'",
-            "form-action *",
-        ].join("; ");
-        res.set_mut((modifiers::Header(StrictTransportSecurity::excluding_subdomains(31536000u64)),
-                     modifiers::Header(ContentSecurityPolicy(csp.clone())),
-                     modifiers::Header(XContentSecurityPolicy(csp)),
-                     modifiers::Header(XContentTypeOptions("nosniff".to_string())),
-                     modifiers::Header(XXSSProtection("1; mode=block".to_string())),
-                     modifiers::Header(XFrameOptions("DENY".to_string()))));
-    }
-}
-impl AfterMiddleware for DefaultHeadersMiddleware {
-    fn after(&self, _: &mut Request, mut res: Response) -> IronResult<Response> {
-        self.set_headers(&mut res);
-        Ok(res)
-    }
-    fn catch(&self, _: &mut Request, mut err: IronError) -> IronResult<Response> {
-        self.set_headers(&mut err.response);
-        Err(err)
     }
 }
