+++ src/middleware/mod.rs
@@ -1,0 +1,5 @@
+mod logging;
+pub use self::logging::LogMiddleware;
+
+mod headers;
+pub use self::headers::DefaultHeadersMiddleware;
